Fixed the "none" names which were given to concepts with no preferred term
Alter the hierarchyBuilder.py's determinePreferred() method by saving
the first english term found. This english term is now used if no
preferred term for the concept is found to avoid the "none" which used
to show up. If no english term either, than a "No_English_Term" label
will be used.

In the UMLSSubsetBuilder.py, changed the querying for names to include
the language (LAT) in addition to the TTY and string (STR).