Updated UMLSSubsetBuilder.py and hierarchybuilder.py for efficiency by only translating the relationships dictionary and others when debugMode is initiated
The program now uses just the cuis, rather than cui-names, throughout
program execution. Rather than translating every dictionary and list
with every excution, as previously, the program now holds off on
translations until the very end. This improves efficiency by preventing
unnecessary querying and datastructure creation from the that used to
occur when not running in the program in debug mode. No changes to the
actual hierarchy structure were made besides a rearrangment of the names
of the entries in the loops dictionary. EFFICIENCY UPDATE.